requirements.txt
torch>=1.13
torchvision
opencv-python
numpy
pandas
tqdm
matplotlib
Pillow
